Compilação da camera ok

+++ Camera.h
@@ -9,7 +9,7 @@ public :
     Camera ();
 
     void setCamera(QVector3D eye, QVector3D at, QVector3D up);
-    void zoom(int zoom);
+    void zoom(double zoom);
 
     QVector3D eye;
     QVector3D at;
